OpenRouter-Monitor/requirements.txt
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
cairosvg>=2.7.0
